+++ test/TensorFlowNET.Keras.UnitTest/Layers/Layers.Reshaping.Test.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using NumSharp;
+using Tensorflow.Numpy;
 using static Tensorflow.Binding;
 using static Tensorflow.KerasApi;
 
@@ -12,7 +12,7 @@ namespace TensorFlowNET.Keras.UnitTest
         public void ZeroPadding2D()
         {
             var input_shape = new[] { 1, 1, 2, 2 };
-            var x = np.arange(np.prod(input_shape)).reshape(input_shape);
+            var x = np.arange((int)np.prod(input_shape)).reshape(input_shape);
             var zero_padding_2d = keras.layers.ZeroPadding2D(new[,] { { 1, 0 }, { 1, 0 } });
             var y = zero_padding_2d.Apply(x);
             Assert.AreEqual((1, 2, 3, 2), y.shape);
@@ -22,7 +22,7 @@ namespace TensorFlowNET.Keras.UnitTest
         public void UpSampling2D()
         {
             var input_shape = new[] { 2, 2, 1, 3 };
-            var x = np.arange(np.prod(input_shape)).reshape(input_shape);
+            var x = np.arange((int)np.prod(input_shape)).reshape(input_shape);
             var y = keras.layers.UpSampling2D(size: (1, 2)).Apply(x);
             Assert.AreEqual((2, 2, 2, 3), y.shape);
         }
